+++ include/utils/Display.h
@@ -15,7 +15,7 @@ public:
 
 	Display(int width, int height, const std::string& title);
 	virtual ~Display();
-	void Clear(float r, float g, float b, float a);
+	void Clear(double r, double g, double b, double a);
 	void SwapBuffers();
 	bool WindowShouldClose();
 	bool Refresh();
